Deleted assest folder and moved images and SVGs to public folder
