+++ index.html
@@ -2,27 +2,9 @@
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
-    <link rel="icon" type="image/svg+xml" href="./src/assets/SVG/logo.svg" />
+    <link rel="icon" type="image/svg+xml" href="./SVG/logo.svg" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Lesedi Molemi</title>
-    <script type="text/javascript">
-      (function (l) {
-        if (l.search[1] === "/") {
-          var decoded = l.search
-            .slice(1)
-            .split("&")
-            .map(function (s) {
-              return s.replace(/~and~/g, "&");
-            })
-            .join("?");
-          window.history.replaceState(
-            null,
-            null,
-            l.pathname.slice(0, -1) + decoded + l.hash
-          );
-        }
-      })(window.location);
-    </script>
   </head>
   <body class="bg-secondary text-white pr-3 pl-3 pb-6">
     <div id="root"></div>
